Refactor cursor handling and viewport management
- Introduced `CursorRenderState` to manage cursor visibility and position.
- Added `FocusedOutputState` to track the currently focused output.
- Implemented `SoftwareCursorCache` for efficient cursor rendering.
- Enhanced viewport management with `initialize_scene_geometry_from_surface` and related functions.
- Updated `window_viewport_projection_system` to handle output-local coordinates.
- Added support for output-bound layer shell surfaces.
- Improved action configuration parsing to support new action types.
- Added tests for overlapping floating windows to ensure correct pointer event routing.
- Introduced `output_workspace_housekeeping_system` to manage workspace assignments for outputs.
- Updated IPC commands to include viewport-related actions.
- Added new shell configuration options for external shell integration.

+++ crates/nekoland-ecs/src/events/window_events.rs
@@ -18,6 +18,6 @@ pub struct WindowClosed {
 #[derive(Message, Clone, Debug, Default, Serialize, Deserialize, PartialEq, Eq)]
 pub struct WindowMoved {
     pub surface_id: u64,
-    pub x: i32,
-    pub y: i32,
+    pub x: i64,
+    pub y: i64,
 }
